In golfcar_ppc/speed_advisor:
- Approaching to a final solution with intersection_handler.
Note: with the current solution, the car stops on the intersection point,
whereas before it stopped a few meters in front. This could be corrected
with an offset, but this offset also makes the positioning of the intersection
point on the map difficult. I must talk to Demian to see what's the best
solution.

- Minor comments in golfcar_purepursuit