In tencon: changed the logging scheme
With the previous logging scheme (log all the simulation to a variable, and
then save that variable to a file), there were some MemoryError issues
when the simulation was too long.

I changed to logging directly to the file at each simulation step. Logging
as ascii, one line at a time. At each logged step (a line), I log the time,
and the id, pos and vel of all mobiles.

Logging each simulation to its own gzip file.