lowlevel arduino
Got the new arduino shield (PCB) by Sethu.
- arduino_node:
  + made it a class
  + using a timer rather than sleep for main loop
- step_pwm_lights:
  + changed the pin numbers

- diagnostic_updater:
there is a problem with get_param('~diagnostic_period'). throwing
exceptions. catching them for now.
I think the cause is that I am calling it too often. In the C++ version
it was a get_param_cached call, which is probably limiting the calling
rate. Check with WG and/or implement my own caching.