Created framework for tracking
It's a ROS package named "tracking" that will become a library that can be
reused by other packages. Implemented in C++ using BFL.

Currently only doing some testings with EKF. Compared to BFL tutorial,
I removed the input and so the EKF has to track the velocity by itself.

Created a Path class that implements a parametric cubic B-spline curve,
useful for testing with a simulated robot that follows a trajectory
defined by some waypoints.