infrastructure_road_monitoring
modified the background class so that the alpha value starts at 1 and decreases
to the target value slowly. This avoids the problem that the first frame had
too much importance.

introduced a reset chanel in background extraction node. Also modified a bit the
topic names.

fixed a problem with the python scripts communicating with the DB server:
after the public IP change with port remapping, the fmautonomy.no-ip.info
solution does not work so well any more. So I changed to using the IP
directly. And to make it easier, I made that the URL must be set in the
parameter server compulsory. Created a function to check that it is, and
test that URL.