speed_advisor
Modified the SpeedAttribute and SpeedSettings classes, principally
so that it can take care of the automode button, and also keep a minimum
velocity when going to stop.

The SpeedSettings object keeps track of the desired velocity (so that it
can generate velocity ramps). When switching from manual driving mode to
autonomous mode, this is set to the current velocity as given by the odometry.

When the target velocity is below 0.4, then instead of ramping down all the way
to 0, maintain a velocity of 0.4, until the ramp reaches 0, then stop abruptly.
This is because the golfcart motors cannot easily follow a very small speed.