Created a simulation launch and world file for the dense map
based on those in local_map_ext, but without rrt related stuffs. Besides,
the bitmap used in stage is based on the dense map, but hand-drawn a bit larger
than the actual dense map, otherwise the car stops all the time (misalignment).
Note: the contour is dotted on purpose, otherwise there is a problem with
selecting items in stage.