changes in chrony.conf on lowlevel
The system clock is initialized with the hwclock at bootup. This clock drifts
hence the initial system time is off. With Chrony and a connection to a time
server it will converge slowly to the correct time.

Chrony has an option to estimate the error of the hardware clock and thus to
init the system time to a more accurate value. This requires the so called
RTC option to be turned on. This is what this update is doing. To be tested.