Improved ntp_monitor
The program had a fault: if ntpdate takes too long to respond (and it usually
does), then the diagnostic message is not sent often enough, resulting in
robot_monitor marking it as stale.

With this fix, I use 2 timers: one to measure the offset that is run every
60 seconds, and one to report the offset (every 500ms). The message contains
information about when the last measure was performed and how long ago this
was.